Taller Final/ecomarket_agent/requirements.txt
# Core
langchain==0.3.4
langchain-core==0.3.12
langchain-openai==0.2.3

# Vector DB and embeddings
chromadb>=0.3.28

# LLM provider
openai>=0.27.0

# Frontend
streamlit>=1.20.0
python-dotenv>=1.0.0

# Document loaders and utils (si necesitas procesar PDFs, CSVs)
pypdf>=3.7.0
pandas>=2.0.0
unstructured>=0.7.0
requests>=2.28.0
